requirements.txt
alembic==1.0.2
aniso8601==4.0.1
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
bcrypt==3.1.6
bokeh==0.13.0
boto==2.49.0
boto3==1.9.43
botocore==1.12.43
bz2file==0.98
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
coverage==4.5.2
cryptography==2.4.1
cycler==0.10.0
decorator==4.3.0
docutils==0.14
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.6
Flask-Migrate==2.3.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
ForgeryPy==0.1
ForgeryPy3==0.3.1
gcloud==0.18.3
gensim==3.6.0
google-api-core==1.5.2
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-speech==0.36.0
googleapis-common-protos==1.5.5
grpcio==1.16.0
gunicorn==19.9.0
httplib2==0.12.0
idna==2.6
ipython==7.0.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jwcrypto==0.6.0
kiwisolver==1.0.1
Mako==1.0.7
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
nltk==3.3
nose==1.3.7
numpy==1.15.4
oauth2-client==1.1.0
oauth2client==4.1.3
packaging==18.0
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
protobuf==3.6.1
psycopg2==2.7.7
ptyprocess==0.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
Pygments==2.2.0
PyJWT==1.7.1
pyparsing==2.2.2
python-dateutil==2.7.3
python-editor==1.0.3
python-jwt==3.2.3
pytz==2018.5
PyYAML==3.13
requests==2.18.4
rsa==4.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
smart-open==1.7.1
SpeechRecognition==3.8.1
SQLAlchemy==1.2.14
tensorboard==1.11.0
termcolor==1.1.0
tornado==5.1.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.14.1
